mypage파일 user로 통일
