--- api/services/mypageService.js
@@ -1,11 +1,0 @@
-const { mypageDao } = require('../models');
-
-const getUserInformation = async (userId) => {
-    return await mypageDao.getUserInformation(userId);
-};
-
-const getUserOrdered = async (userId) => {
-    return await mypageDao.getUserOrdered(userId);
-};
-
-module.exports = { getUserInformation, getUserOrdered };
